--- src/mobile_app/lib/account_setup/blocs/login/login_bloc.dart
@@ -1,18 +1,0 @@
-import 'dart:async';
-
-import 'package:bloc/bloc.dart';
-import 'package:equatable/equatable.dart';
-
-part 'login_event.dart';
-part 'login_state.dart';
-
-class LoginBloc extends Bloc<LoginEvent, LoginState> {
-  LoginBloc() : super(LoginInitial());
-
-  @override
-  Stream<LoginState> mapEventToState(
-    LoginEvent event,
-  ) async* {
-    // TODO: implement mapEventToState
-  }
-}
