--- src/mobile_app/lib/account_setup/blocs/login/login_event.dart
@@ -1,5 +1,0 @@
-part of 'login_bloc.dart';
-
-abstract class LoginEvent extends Equatable {
-  const LoginEvent();
-}
