--- src/mobile_app/lib/account_setup/blocs/login/login_state.dart
@@ -1,10 +1,0 @@
-part of 'login_bloc.dart';
-
-abstract class LoginState extends Equatable {
-  const LoginState();
-}
-
-class LoginInitial extends LoginState {
-  @override
-  List<Object> get props => [];
-}
